refactor event
